requirements.txt
numpy>=1.21,<2.0
opencv-python>=4.5,<5.0
albumentations==0.4.3
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.0.8
omegaconf==2.0.0
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
more-itertools>=8.0.0
transformers==4.37.2
timm==0.9.7
pillow==9.4.0
tqdm==4.67.1
diffusers>=0.19,<0.30
tensorboard==2.13.0
huggingface_hub==0.33.0
open_clip_torch==2.22.0
matplotlib==3.7.5
flash-attn==2.8.3
git+https://github.com/openai/CLIP.git@main